worker/requirements.txt
nats-py>=2.6.0
pyone==6.10.0
redis>=5.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
